restructure
unbranching noise3 noise2 files back into the single noise file
swapping abs and pow in noiseprocessing to std::abs and std::pow
disabling internal testing in point-cloud example
