+++ examples/histogram/histogram.cpp
@@ -1,4 +1,4 @@
-#include <noiz2/noise2.hpp>
+#include <noiz/noise2.hpp>
 #include <charconv>
 #include <filesystem>
 #include <format>
