+++ examples/texture2d/texture2d.cpp
@@ -1,4 +1,4 @@
-#include <noiz2/noise2.hpp>
+#include <noiz/noise2.hpp>
 
 #include <argument_parsing.hpp>
