+++ noiz/include/noiz/detail/grid2.hpp
