+++ noiz/include/noiz/noise-processing3.hpp
@@ -1,16 +1,17 @@
 #pragma once
-#include "noise2.hpp"
+#include "noise3.hpp"
+
 
 namespace noiz{
 //static functions, maybe make members of noise2? not really sure
 template <std::floating_point Type>
-class Noise_Processor2 {
+class Noise_Processor3 {
 public:
-	explicit Noise_Processor2(noiz::Noise2<Type>& noise) : noise{noise}{}
+	explicit Noise_Processor3(noiz::Noise3<Type>& noise) : noise{noise}{}
 	//if this is multi-threaded, accessing these variables directly could lead to strange behavior
 	//i.e. a handful of threads processing noise, then another comes in and changes octave??
 	//not sure if thats a use case worth covering?
-	noiz::Noise2<Type>& noise;
+	noiz::Noise3<Type>& noise;
 
 	//default values
 	uint8_t octave{ 6 };
@@ -25,14 +26,12 @@ public:
 	Type hbf_offset = (Type)0.7;
 	std::vector<Type> hbf_exponent_array;
 
-	//could construct an object of noiz::noise2 here and it would simplify these functions a little bit.
-
-	auto raw_noise(noiz::Vec2<Type> const& point) -> Type{
+	auto raw_noise(noiz::Vec3<Type> const& point) -> Type{
 		//redundant
 		return noise.at(point * step);
 	}
 
-	auto basic_processing(noiz::Vec2<Type> const& point) -> Type {
+	auto basic_processing(noiz::Vec3<Type> const& point) -> Type {
 		
 		Type total = Type(0);
 		Type frequency = (Type)2;
@@ -47,14 +46,14 @@ public:
 		return total/normalizer;
 	}
 
-	auto turbulence_processing(noiz::Vec2<Type> const& point) -> Type {
+	auto turbulence_processing(noiz::Vec3<Type> const& point) -> Type {
 		Type amplitude = this->amplitude;
 		Type frequency = this->frequency;
 		
 			Type sum = 0.f;
 			Type normalizer = 0.f;
 			for (int i = 0; i < octave; i++) {
-				sum += noise.at(noiz::Vec2<Type>{.x = point.x * frequency, .y = point.y * frequency}) * amplitude;
+				sum += noise.at(point * frequency) * amplitude;
 				normalizer += amplitude;
 				amplitude *= persistence;
 				frequency *= lacunarity;
@@ -64,26 +63,25 @@ public:
 			return sum;
 	}
 
-	auto billowy_processing(noiz::Vec2<Type> const& point) -> Type {
-		return abs(noise.at(point));
+	auto billowy_processing(noiz::Vec3<Type> const& point) -> Type {
+		return std::abs(noise.at(point));
 	}
 
-	auto rigid_processing(noiz::Vec2<Type> const& point) -> Type {
+	auto rigid_processing(noiz::Vec3<Type> const& point) -> Type {
 		return 1.f - std::abs(noise.at(point));
 	}
 
-	auto hybrid_multi_fractal_processing(noiz::Vec2<Type> const& point) -> Type {
+	auto hybrid_multi_fractal_processing(noiz::Vec3<Type> const& point) -> Type {
 		//https://www.classes.cs.uchicago.edu/archive/2015/fall/23700-1/final-project/MusgraveTerrain00.pdf
 		
-		//this function assumes the octave could be a decimal value
-		//i also casted it down from 3d to 2d, im not sure what effects this has. hard to say without 3d rendering
+		//this function assumes the octave could be a floating point value
 		
 		//double HybridMultifractal( Vector point, double H, double lacunarity,
 		//double octaves, double offset )
-		noiz::Vec2<Type> tempPoint = point;
+		noiz::Vec3<Type> tempPoint = point;
 
-		Type frequency, result, signal, weight, remainder;
-		//double Noise3(); what is this??
+		Type frequency, result, signal, weight;//, remainder; if octave is floating
+
 		/* precompute and store spectral weights */
 		if (hbf_exponent_array.size() != octave) {
 			hbf_exponent_array.resize(octave);
@@ -92,7 +90,7 @@ public:
 			frequency = 1.0;
 			for (int i = 0; i < octave; i++) {
 				/* compute weight for each frequency */
-				hbf_exponent_array[i] = pow(frequency, -hbf_H);
+				hbf_exponent_array[i] = std::pow(frequency, -hbf_H);
 				frequency *= lacunarity;
 			}
 		}
@@ -104,7 +102,7 @@ public:
 		/* spectral construction inner loop, where the fractal is built */
 		for (int i = 1; i < octave; i++) {
 			/* prevent divergence */
-			if (weight > 1.0) weight = 1.0;
+			if (weight > 1.0) {weight = 1.0;}
 			/* get next higher frequency */
 			signal = (noise.at(tempPoint) + hbf_offset) * hbf_exponent_array[i];
 			/* add it in, weighted by previous freq's local value */
@@ -114,34 +112,23 @@ public:
 			weight *= signal;
 			/* increase frequency */
 			tempPoint = tempPoint * lacunarity;
-		} /* for */
-		/* take care of remainder in “octaves” */
+		}
 		
 		/* octave is currently an int, may swap for this
+		* take care of remainder in “octaves”
 		remainder = octave - (int)octave;
 		if (remainder)
 			// “i” and spatial freq. are preset in loop above
 			result += remainder * noise.at(point) * hbf_exponent_array[octave - 1];
 		*/
 
-	/* value checking
-		static Type biggestValue = (Type)0;
-		static Type smallestValue = (Type)1;
-		if (result > biggestValue) {
-			biggestValue = result;
-			std::cout << "biggest result : " << result << std::endl;
-		}
-		else if (result < smallestValue) {
-			smallestValue = result;
-			std::cout << "smallest result : " << result << std::endl;
-		}
-	*/
-
+	
+		//values from 0 to ~4 can be expected. highest value ive seen is 3.7
 		return result;
 	}
 
 /*
-	auto blended_basic_processing(noiz::Vec2<Type> const& point, std::vector<Noise2<Type>> const& noise){
+	auto blended_basic_processing(noiz::Vec3<Type> const& point, std::vector<Noise3<Type>> const& noise){
 		*each noise source would need its own set of variables
 
 		std::vector<octaves>
