+++ tests/tests/test_data2.cpp
@@ -1,4 +1,4 @@
-#include <noiz2/detail/data2.hpp>
+#include <noiz/detail/data2.hpp>
 #include <test/test.hpp>
 
 namespace {
